+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v2.5.6 (FUTURE)
 
 * [#2839](https://github.com/digitalocean/netbox/issues/2839) - Add "110 punch" type for pass-through ports
 * [#2854](https://github.com/digitalocean/netbox/issues/2854) - Enable bulk editing of pass-through ports
+* [#2866](https://github.com/digitalocean/netbox/issues/2866) - Add cellular interface types (GSM/CDMA/LTE)
 
 ## Bug Fixes